requirements.txt
fastapi
uvicorn
sse-starlette
python-dotenv
openai
anthropic
aisuite
redis
click
docstring-parser
